% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.dem.R
\name{generate.dem}
\alias{generate.dem}
\title{turn elevation raster .tif files into a DEM}
\usage{
generate.dem(
  map.data,
  raster.files,
  coordinate.system = NULL,
  aggregate = NULL
)
}
\arguments{
\item{map.data}{An sf data frame with geometries of type POLYGON or MULTIPOLYGON and associated attributes.}

\item{raster.files}{A vector of .tif elevation raster file names you wish to turn into a DEM. If empty the function will pull raster data for the map.data bounding box from AWS Terrain Tiles.}

\item{coordinate.system}{EPSG code of the projection you wish to use. Default NULL: if crs of map.data and the generated DEM do not match the latter will be projected to the crs of the former.}

\item{aggregate}{factor to aggregate the generated DEM by. Default NULL.}
}
\value{
A raster object.
}
\description{
generate.dem converts elevation raster .tif files into a DEM ready to use with overlay.relief
}
