% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.map.R
\name{generate.map}
\alias{generate.map}
\title{generate basic relief maps}
\usage{
generate.map(data, x, y, hillshade, variable, coordinate.system)
}
\arguments{
\item{data}{output of overlay.relief}

\item{x}{name of the longitude variable in the overlay.relief output as character vector}

\item{y}{name of the latitude variable in the overlay.relief output as character vector}

\item{hillshade}{name of the hillshade variable in the overlay.relief output as character vector}

\item{variable}{name of the variable in the overlay.relief output you wish to map as character vector}

\item{coordinate.system}{EPSG code of the projection used in overlay.relief}
}
\value{
ggplot object
}
\description{
generate.map generates a ggplot map visualizing sf attributes over shaded relief.
It gives you a basic map so you don't have to deal with the nitty gritty and can just
focus on customizing the aesthetic.
}
