% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.relief.R
\name{overlay.relief}
\alias{overlay.relief}
\title{overlay relief with coloured data visualization}
\usage{
overlay.relief(
  map.data,
  variables,
  elevation.raster,
  make.hillshade = TRUE,
  coordinate.system = NULL,
  altitude = 45,
  azimuth = 270,
  z.factor = NULL
)
}
\arguments{
\item{map.data}{An sf data frame with geometries of type POLYGON or MULTIPOLYGON and associated attributes.}

\item{variables}{The column names of the attributes you wish to visualize as a character vector i.e c("variable_1","variable_2").}

\item{elevation.raster}{Digital Elevation Model (if make.hillshade = TRUE) or Shaded Relief (if make.hollshade = FALSE) data of class RasterLayer.}

\item{make.hillshade}{Set to TRUE if you wish to turn DEM data into hillshade, set to FALSE if you wish to pass a hillshade raster to the function directly. Default TRUE.}

\item{coordinate.system}{EPSG code of the projection you wish to use. Default NULL: if crs of map.data and elevation.raster do not match the latter will be projected to the crs of the former.}

\item{altitude}{Elevation angle of light source in degrees to calculate hill shade. Numeric value between 0 and 90.}

\item{azimuth}{Direction angle of light source in degrees to calculate hill shade. Numeric value between 0 and 360.}

\item{z.factor}{Numeric value to multiply elevation raster by to exaggerate relief. Default NULL.}
}
\value{
A data frame of points and their associated hill shade values and sf attributes.
}
\description{
overlay.relief converts DEM data into a data frame of spatial points and their associated
hill shade values. It merges this data with attributes associated with sf geometries. This allows
you to visualize sf attributes on top of a shaded relief in ggplot without masking the relief.
}
