% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.relief.R
\name{overlay.relief}
\alias{overlay.relief}
\title{overlay relief with coloured data visualization}
\usage{
overlay.relief(
  map.data,
  variables,
  elevation.raster,
  coordinate.system = 4326,
  altitude = 45,
  azimuth = 270,
  z.factor = NULL
)
}
\arguments{
\item{map.data}{An sf data frame with geometries of type MULTILINESTRING, POLYGON or MULTIPOLYGON and associated attributes.}

\item{variables}{The column names of the attributes you wish to visualize on top of the relief passed to the argument as a character vector i.e c("variable_1","variable_2").}

\item{elevation.raster}{Digital Elevation Model data in form of an object of class RasterLayer.}

\item{coordinate.system}{Four digit epsg code of the projection you wish to use. Default 4326.}

\item{altitude}{Elevation angle of light source in degrees to calculate hill shade. Numeric value between 0 and 90.}

\item{azimuth}{Direction angle of light source in degrees to calculate hill shade. Numeric value between 0 and 360.}

\item{z.factor}{Numeric value to multiply elevation raster by to exaggerate relief. Default NULL.}
}
\value{
A data frame of points and their associated hill shade values and sf attributes.
}
\description{
This function converts DEM data into a data frame of spatial points and their associated
hill shade values. It merges this data with atributes associated with sf geometries. This allows
you to visualize sf attributes on top of a shaded relief in ggplot without masking the relief.
To implement this in ggplot plot the function output as a geom_raster() with x and y as point coordinates.
Control the hill shading with an alpha scale and the sf attribute visualization with a fill scale.
}
